reclaim_ws/src/reclaim_control/scripts/requirements.txt
pyserial>=3.5,<4.0
PyYAML>=6.0,<7.0
